Change the way to format the RDF removing unnecessary code and changing how to write into the buffer, this has as result the following.

command: go test -bench . -benchtime 10s

// 1 item (string value) x operation
BenchmarkToRDF_New-4            100000000              201 ns/op               4 B/op          1 allocs/op
BenchmarkToRDF_Current-4        20000000               890 ns/op             136 B/op         12 allocs/op

// 4 items (string, password, int values) x operation
BenchmarkToRDF_New-4            10000000              1762 ns/op             272 B/op         16 allocs/op
BenchmarkToRDF_Current-4         3000000              5318 ns/op             936 B/op         69 allocs/op

// 5 items (1 Geo value) x operation
BenchmarkToRDF_New-4             3000000              5872 ns/op            1264 B/op         38 allocs/op
BenchmarkToRDF_Current-4         1000000             10452 ns/op            2136 B/op        104 allocs/op

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/dgraph-io/dgraph/650)
<!-- Reviewable:end -->