@ashishnegi 2017-03-02
Thanks for the contribution. Looks good improvement. :)
I have a few comments. Please have a look. 
Thanks again.

---

Review status: 0 of 2 files reviewed at latest revision, 10 unresolved discussions.

---

*[worker/backup.go, line 31 at r1](https://reviewable.io:443/reviews/dgraph-io/dgraph/650#-KeCGVpkKvkl-UTLnCIE:-KeCGVplWEr_Fv0ywS5R:b-yqibwe) ([raw file](https://github.com/dgraph-io/dgraph/blob/39189970ed65749c0d800d8bac0ab051a5cf466d/worker/backup.go#L31)):*
> ```Go
> 	pl := item.list
> 	for _, p := range pl.Postings {
> 		x.Check2(buf.WriteString(item.prefix))
> ```

have you removed the check knowing that buf.WriteString always returns nil ?

---

*[worker/backup.go, line 43 at r1](https://reviewable.io:443/reviews/dgraph-io/dgraph/650#-KeCGcj3WysFpf7JKCxF:-KeCGcj3WysFpf7JKCxG:b-dnpmyw) ([raw file](https://github.com/dgraph-io/dgraph/blob/ce87bcb8991678685bcfabacb9b94159c3b13230/worker/backup.go#L43)):*
> ```Go
> 			x.Check(err)
> 			stringValue := str.Value.(string)
> 			buf.WriteString("\"")
> ```

WriteRune when there is simple rune ..

---

*[worker/backup.go, line 45 at r1](https://reviewable.io:443/reviews/dgraph-io/dgraph/650#-KeCH3YZxc1Y71R6734E:-KeCH3YZxc1Y71R6734F:bx9ip1w) ([raw file](https://github.com/dgraph-io/dgraph/blob/ce87bcb8991678685bcfabacb9b94159c3b13230/worker/backup.go#L45)):*
> ```Go
> 			buf.WriteString("\"")
> 			buf.WriteString(stringValue)
> 			buf.WriteString("\"")
> ```

writeRune

---

*[worker/backup.go, line 51 at r1](https://reviewable.io:443/reviews/dgraph-io/dgraph/650#-KeCGiY72zQcCwKfWv4s:-KeCGiY72zQcCwKfWv4t:b-4mhdcc) ([raw file](https://github.com/dgraph-io/dgraph/blob/ce87bcb8991678685bcfabacb9b94159c3b13230/worker/backup.go#L51)):*
> ```Go
> 				buf.WriteString("^^<pwd:")
> 				buf.WriteString(vID.Name())
> 				buf.WriteString(">")
> ```

WriteRune

---

*[worker/backup.go, line 56 at r1](https://reviewable.io:443/reviews/dgraph-io/dgraph/650#-KeCGlzS9ThWi3sQvfOP:-KeCGlzT1VfQQVJDe8hR:b-4mhdcc) ([raw file](https://github.com/dgraph-io/dgraph/blob/ce87bcb8991678685bcfabacb9b94159c3b13230/worker/backup.go#L56)):*
> ```Go
> 				buf.WriteString("^^<xs:")
> 				buf.WriteString(vID.Name())
> 				buf.WriteString(">")
> ```

WriteRune

---

*[worker/backup.go, line 58 at r1](https://reviewable.io:443/reviews/dgraph-io/dgraph/650#-KeCGo5_4YZylzJFYuUo:-KeCGo5aXmfKLsa9feu4:b-4mhdcc) ([raw file](https://github.com/dgraph-io/dgraph/blob/ce87bcb8991678685bcfabacb9b94159c3b13230/worker/backup.go#L58)):*
> ```Go
> 				buf.WriteString(">")
> 			} else if len(p.Lang) > 0 {
> 				buf.WriteString("@")
> ```

WriteRune

---

*[worker/backup.go, line 64 at r1](https://reviewable.io:443/reviews/dgraph-io/dgraph/650#-KeCGqLbHZbLjgno3OGC:-KeCGqLbHZbLjgno3OGD:bjzyapq) ([raw file](https://github.com/dgraph-io/dgraph/blob/ce87bcb8991678685bcfabacb9b94159c3b13230/worker/backup.go#L64)):*
> ```Go
> 			buf.WriteString("<0x")
> 			buf.WriteString(strconv.FormatUint(p.Uid, 16))
> 			buf.WriteString(">")
> ```

WriteRune and similarly below as well.

---

*[worker/backup_test.go, line 270 at r1](https://reviewable.io:443/reviews/dgraph-io/dgraph/650#-KeCJ8RPzAoCwpoLl69K:-KeCJ8RPzAoCwpoLl69L:b-y0lc7s) ([raw file](https://github.com/dgraph-io/dgraph/blob/ce87bcb8991678685bcfabacb9b94159c3b13230/worker/backup_test.go#L270)):*
> ```Go
> }
> 
> func BenchmarkToRDF_New(b *testing.B) {
> ```

`_New` is not required.

---

*[worker/backup_test.go, line 274 at r1](https://reviewable.io:443/reviews/dgraph-io/dgraph/650#-KeCItaMX1YguSJcuOhv:-KeCItaMX1YguSJcuOhw:b-vibzyv) ([raw file](https://github.com/dgraph-io/dgraph/blob/ce87bcb8991678685bcfabacb9b94159c3b13230/worker/backup_test.go#L274)):*
> ```Go
> }
> 
> func runBenchToRDF(b *testing.B, f func(buf *bytes.Buffer, item kv)) {
> ```

will be better to have the code here be part of `BenchmarkToRDF_New`

---

*[worker/backup_test.go, line 280 at r1](https://reviewable.io:443/reviews/dgraph-io/dgraph/650#-KeCIhxtiduKrUhIxtGd:-KeCIhxtiduKrUhIxtGe:bfcm7qf) ([raw file](https://github.com/dgraph-io/dgraph/blob/ce87bcb8991678685bcfabacb9b94159c3b13230/worker/backup_test.go#L280)):*
> ```Go
> 	items := generateBenchValues()
> 
> 	var i int
> ```

not needed.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/dgraph-io/dgraph/650#-:-KeCJCrp30loorK3IyZ3:b-noztjp)*
<!-- Sent from Reviewable.io -->